agent/requirements.txt
# Stage B
discord.py==2.3.2
tomli==2.0.1
psutil==5.9.8
numpy<2.0          # 解决 Numpy 2.0 C-API 不兼容问题
pandas==2.0.3      # 锁定旧版以兼容 AlmaLinux 8 / Python 3.9 编译环境
finnhub-python==2.4.20  # 数据引擎 Finnhub 源（备用 / 降级）
tomli-w==1.0.0
requests>=2.31.0
curl_cffi==0.13.0
trafilatura>=1.6.0
